+++ transports/openai-realtime-webrtc-transport/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to **Pipecat OpenAIRealTimeWebRTCTransport** will be documen
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Added default case to `sendMessage` to properly throw an `UnsupportedFeatureError`
+
 ## [1.4.0]
 
 - Bump client-js version to work with latest 1.5.0 and support latest features